requirements.txt
python-telegram-bot>=21.0
google-genai>=1.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
Pillow>=10.0.0
aiohttp>=3.9.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
